requirements.txt
pymongo 
pandas 
numpy 
pyYAML
dill 
python-dotenv
scikit-learn
-e .